+++ aries-site/src/examples/foundation/accessibility/index.js
@@ -1,1 +1,1 @@
-export * from './GuideCards';
+export * from './SectionCards';
